colocate the daemon test
