+++ tests/core/daemon/test_daemon.py
@@ -3,31 +3,21 @@ import asyncio
 import json
 import logging
 import pytest
-import signal
-import sys
-import time
-
 
 from dataclasses import dataclass, replace
 from typing import Any, Dict, List, Optional, Type, Union, cast
 
 from chia.daemon.keychain_server import DeleteLabelRequest, SetLabelRequest
-from chia.daemon.client import connect_to_daemon_and_validate
 from chia.daemon.server import WebSocketServer, service_plotter
 from chia.server.outbound_message import NodeType
 from chia.types.peer_info import PeerInfo
 from chia.util.ints import uint16
 from chia.util.keychain import KeyData
 from chia.daemon.keychain_server import GetKeyRequest, GetKeyResponse, GetKeysResponse
-from chia.util.config import lock_and_load_config, save_config
 from chia.util.keyring_wrapper import DEFAULT_PASSPHRASE_IF_NO_MASTER_PASSPHRASE
-from chia.util.misc import termination_signals
 from chia.util.ws_message import create_payload
 from tests.core.node_height import node_height_at_least
-from chia.simulator.socket import find_available_listen_port
 from chia.simulator.time_out_assert import time_out_assert_custom_interval, time_out_assert
-from tests.core.data_layer.util import ChiaRoot
-from tests.util.misc import closing_chia_root_popen
 
 
 # Simple class that responds to a poll() call used by WebSocketServer.is_running()
@@ -762,31 +752,3 @@ async def test_key_label_methods(
     keychain.add_private_key(test_key_data.mnemonic_str(), "key_0")
     await ws.send_str(create_payload(method, parameter, "test", "daemon"))
     assert_response(await ws.receive(), response_data_dict)
-
-
-@pytest.mark.parametrize(argnames="signal_number", argvalues=termination_signals)
-@pytest.mark.asyncio
-async def test_daemon_terminates(signal_number: signal.Signals, chia_root: ChiaRoot) -> None:
-    port = find_available_listen_port()
-    with lock_and_load_config(root_path=chia_root.path, filename="config.yaml") as config:
-        config["daemon_port"] = port
-        save_config(root_path=chia_root.path, filename="config.yaml", config_data=config)
-
-    with closing_chia_root_popen(chia_root=chia_root, args=[sys.executable, "-m", "chia.daemon.server"]) as process:
-        start = time.monotonic()
-        while time.monotonic() - start < 15:
-            client = await connect_to_daemon_and_validate(root_path=chia_root.path, config=config)
-            if client is not None:
-                break
-            await asyncio.sleep(0.1)
-        else:
-            raise Exception("unable to connect")
-
-        try:
-            return_code = process.poll()
-            assert return_code is None
-
-            process.send_signal(signal_number)
-            process.communicate(timeout=5)
-        finally:
-            await client.close()
